refactor: Board에서 점수계산하는 로직 분리
